third_party/blink/web_tests/wpt_internal/overscroll-gestures/toggle-overscroll-command.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" href="mailto:masonf@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<style>
  html, body {
    margin: 0;
  }
  #container {
    width: 200px;
    height: 200px;
    overflow: auto;
  }
  #menu {
    width: 100%;
    height: 100%;
    right: 100%; /* Overscrolls to the left of the container. */
  }
  #content {
    height: 600px;
  }
</style>

<div id="container" overscrollcontainer>
  <div id="menu"></div>
  <div id="content"></div>
</div>
<button id=button command="toggle-overscroll" commandfor="menu"></button>

<script>
promise_test(async () => {
  // The container is only vertically scrollable
  assert_equals(container.scrollWidth, container.clientWidth);
  assert_greater_than(container.scrollHeight, container.clientHeight);

  // Initially the menu is offscreen to the left (and abspos).
  assert_equals(menu.getBoundingClientRect().right, 0);
  assert_not_equals(menu.getBoundingClientRect().left, 0);
  assert_equals(getComputedStyle(menu).position, "absolute");
  assert_not_equals(getComputedStyle(content).position, "absolute");

  // Click the button to bring the menu into view.
  await test_driver.click(button);

  // The menu should now be in view.
  assert_equals(menu.getBoundingClientRect().left, 0);

  // The container has not scrolled.
  assert_equals(container.offsetLeft, 0);

  // Change commandfor, which should remove the overscroll area.
  button.setAttribute("command", "show-popover");
  assert_not_equals(getComputedStyle(menu).position, "absolute", "menu should no longer be an overscroll area");
}, "Activating the toggle-overscroll command invoker brings the overscroll area into view");
</script>
